modified cast type

+++ ft_putnbr_fd.c
@@ -16,7 +16,7 @@ void	ft_putnbr_fd(int n, int fd)
 {
 	long long	ln;
 
-	ln = (long)n;
+	ln = (long long)n;
 	if (ln < 0)
 	{
 		ft_putchar_fd('-', fd);
